# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation using Swagger",
        "version": "1.0.0"
    },
    "paths": {
        "/api/clientes/{cpf}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Consulta cliente por CPF",
                "description": "Retorna os dados do cliente a partir do CPF",
                "operationId": "getClienteByCpf",
                "parameters": [
                    {
                        "name": "cpf",
                        "in": "path",
                        "description": "CPF do cliente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            }
        },
        "/api/cliente": {
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Cadastrar um cliente",
                "description": "Cadastra um novo cliente no sistema",
                "operationId": "createCliente",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cliente"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente cadastrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Retrieve list of users",
                "description": "Returns all users from the database",
                "operationId": "getUsers",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Cliente": {
                "title": "Cliente",
                "description": "Esquema do cliente para a API",
                "required": [
                    "cpf",
                    "nome"
                ],
                "properties": {
                    "id_cliente": {
                        "description": "ID único do cliente",
                        "type": "integer",
                        "nullable": true
                    },
                    "cpf": {
                        "description": "CPF do cliente",
                        "type": "string"
                    },
                    "nome": {
                        "description": "Nome completo do cliente",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "User name",
                        "type": "string"
                    },
                    "email": {
                        "description": "User email",
                        "type": "string",
                        "format": "email"
                    }
                },
                "type": "object"
            }
        }
    }
}